Refactor site generator package
